v0.7.2   * added and updated some new commodity names   * update to new structure of EDDN marketdata dump (listings.csv)   * credits in log can be now more than 2,147,483,647

+++ ED-IBE/IBECompanion/CompanionData.cs
@@ -556,14 +556,15 @@ namespace IBE.IBECompanion
         /// save means, returns "0" if somethings wrong
         /// </summary>
         /// <returns></returns>
-        public Int32 SGetCreditsTotal()
+        public Int64 SGetCreditsTotal()
         {
-            Int32 creditsTotal = 0;
+            Int64 creditsTotal = 0;
 
             try
             {
                 if (CompanionStatus == EDCompanionAPI.Models.LoginStatus.Ok)
-                    creditsTotal = GetValue<Int32>("profile.commander.credits");
+                    creditsTotal = GetValue<Int64>("profile.commander.credits");
+
             }
             catch (Exception)
             {
